Move QgsField data to private header
