+++ src/core/CMakeLists.txt
@@ -504,6 +504,7 @@ SET(QGIS_CORE_HDRS
   qgsfeaturerequest.h
   qgsfeaturestore.h
   qgsfield.h
+  qgsfield_p.h
   qgsfontutils.h
   qgsgeometry.h
   qgsgeometrycache.h
